Reuse conv helpers
