+++ inference-engine/src/gna_plugin/runtime/cnn.cpp
@@ -12,7 +12,9 @@
 #include "backend/dnn_types.h"
 #include "backend/gna_limitations.hpp"
 #include "gna_lib_ver_selector.hpp"
+#include "layers/gna_convolution_layer.hpp"
 
+using namespace GNAPluginNS::GNAConvolutionLayer;
 
 void CNNFilter32(intel_dnn_component_t *component) {
     auto filters = reinterpret_cast<float *>(component->op.conv1D.ptr_filters);
@@ -23,8 +25,7 @@ void CNNFilter32(intel_dnn_component_t *component) {
     const auto convolutionStride = component->op.conv1D.num_feature_map_columns;
     const auto filterSize = component->op.conv1D.num_filter_coefficients;
     const auto numberOfInputs = component->num_columns_in;
-    // TODO: reuse outputFromConv() from backend\am_intel_dnn.cpp
-    const auto numberOfOutputsPerFilter = (numberOfInputs - filterSize) / convolutionStride + 1;
+    const auto numberOfOutputsPerFilter = outputFromConv(numberOfInputs, filterSize, convolutionStride);
     const auto numberOfFilters = component->op.conv1D.num_filters;
 
     std::string layer_name;
